requirements.txt
asgiref==3.3.1
Django==3.1.3
django-phonenumber-field==5.0.0
phonenumbers==8.12.13
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
